Trait sourceview5::prelude::CompletionExt

pub trait CompletionExt: 'static {
    fn add_provider<P: IsA<CompletionProvider>>(&self, provider: &P);

    fn block_interactive(&self);

    fn get_buffer(&self) -> Option<Buffer>;

    fn get_page_size(&self) -> u32;

    fn get_view(&self) -> Option<View>;

    fn hide(&self);

    fn remove_provider<P: IsA<CompletionProvider>>(&self, provider: &P);

    fn set_page_size(&self, page_size: u32);

    fn show(&self);

    fn unblock_interactive(&self);

    fn get_property_remember_info_visibility(&self) -> bool;

    fn set_property_remember_info_visibility(
        &self, 
        remember_info_visibility: bool
    );

    fn get_property_select_on_show(&self) -> bool;

    fn set_property_select_on_show(&self, select_on_show: bool);

    fn get_property_show_icons(&self) -> bool;

    fn set_property_show_icons(&self, show_icons: bool);

    fn connect_hide<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn emit_hide(&self);

    fn connect_provider_added<F: Fn(&Self, &CompletionProvider) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_provider_removed<F: Fn(&Self, &CompletionProvider) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_show<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn emit_show(&self);

    fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_page_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_remember_info_visibility_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_select_on_show_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_show_icons_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Completion methods.

Implementors

Completion

Required methods

fn add_provider<P: IsA<CompletionProvider>>(&self, provider: &P)

Adds an CompletionProvider to the list of providers to be queried for completion results.

Feature: v5_0

provider

an CompletionProvider

fn block_interactive(&self)

fn get_buffer(&self) -> Option<Buffer>

Gets the connected View's Buffer

Feature: v5_0

Returns

A Buffer

fn get_page_size(&self) -> u32

fn get_view(&self) -> Option<View>

Gets the View that owns the Completion.

Feature: v5_0

Returns

A View

fn hide(&self)

Emits the "hide" signal.

When the "hide" signal is emitted, the completion window will be dismissed.

Feature: v5_0

fn remove_provider<P: IsA<CompletionProvider>>(&self, provider: &P)

Removes an CompletionProvider previously added with CompletionExt::add_provider.

Feature: v5_0

provider

an CompletionProvider

fn set_page_size(&self, page_size: u32)

fn show(&self)

Emits the "show" signal.

When the "show" signal is emitted, the completion window will be displayed if there are any results available.

Feature: v5_0

fn unblock_interactive(&self)

fn get_property_remember_info_visibility(&self) -> bool

Determines whether the visibility of the info window should be saved when the completion is hidden, and restored when the completion is shown again.

fn set_property_remember_info_visibility(&self, remember_info_visibility: bool)

Determines whether the visibility of the info window should be saved when the completion is hidden, and restored when the completion is shown again.

fn get_property_select_on_show(&self) -> bool

Determines whether the first proposal should be selected when the completion is first shown.

fn set_property_select_on_show(&self, select_on_show: bool)

Determines whether the first proposal should be selected when the completion is first shown.

fn get_property_show_icons(&self) -> bool

The "show-icons" property denotes if icons should be displayed within the list of completions presented to the user.

fn set_property_show_icons(&self, show_icons: bool)

The "show-icons" property denotes if icons should be displayed within the list of completions presented to the user.

fn connect_hide<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

The "hide" signal is emitted when the completion window should be hidden.

Feature: v5_0

fn emit_hide(&self)

fn connect_provider_added<F: Fn(&Self, &CompletionProvider) + 'static>(
&self,
f: F
) -> SignalHandlerId

The "provided-added" signal is emitted when a new provider is added to the completion.

Feature: v5_0

provider

an CompletionProvider

fn connect_provider_removed<F: Fn(&Self, &CompletionProvider) + 'static>(
&self,
f: F
) -> SignalHandlerId

The "provided-removed" signal is emitted when a provider has been removed from the completion.

Feature: v5_0

provider

an CompletionProvider

fn connect_show<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

The "show" signal is emitted when the completion window should be shown.

Feature: v5_0

fn emit_show(&self)

fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_page_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_remember_info_visibility_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_select_on_show_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_show_icons_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...