Trait sourceview5::prelude::FileSaverExt

pub trait FileSaverExt: 'static {
    fn get_buffer(&self) -> Option<Buffer>;

    fn get_compression_type(&self) -> CompressionType;

    fn get_encoding(&self) -> Option<Encoding>;

    fn get_file(&self) -> Option<File>;

    fn get_flags(&self) -> FileSaverFlags;

    fn get_location(&self) -> Option<File>;

    fn get_newline_type(&self) -> NewlineType;

    fn set_compression_type(&self, compression_type: CompressionType);

    fn set_encoding(&self, encoding: Option<&Encoding>);

    fn set_flags(&self, flags: FileSaverFlags);

    fn set_newline_type(&self, newline_type: NewlineType);

    fn connect_property_compression_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_encoding_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_flags_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_newline_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FileSaver methods.

Implementors

FileSaver

Required methods

fn get_buffer(&self) -> Option<Buffer>

Returns

the Buffer to save.

fn get_compression_type(&self) -> CompressionType

Returns

the compression type.

fn get_encoding(&self) -> Option<Encoding>

Returns

the encoding.

fn get_file(&self) -> Option<File>

Returns

the File.

fn get_flags(&self) -> FileSaverFlags

Returns

the flags.

fn get_location(&self) -> Option<File>

Returns

the gio::File where to save the buffer to.

fn get_newline_type(&self) -> NewlineType

Returns

the newline type.

fn set_compression_type(&self, compression_type: CompressionType)

Sets the compression type. By default the compression type is taken from the File.

compression_type

the new compression type.

fn set_encoding(&self, encoding: Option<&Encoding>)

Sets the encoding. If encoding is None, the UTF-8 encoding will be set. By default the encoding is taken from the File.

encoding

the new encoding, or None for UTF-8.

fn set_flags(&self, flags: FileSaverFlags)

flags

the new flags.

fn set_newline_type(&self, newline_type: NewlineType)

Sets the newline type. By default the newline type is taken from the File.

newline_type

the new newline type.

fn connect_property_compression_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_encoding_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_flags_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_newline_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...