Trait sourceview5::prelude::GutterLinesExt

pub trait GutterLinesExt: 'static {
    pub fn add_class(&self, line: u32, name: &str);

    pub fn add_qclass(&self, line: u32, qname: Quark);

    pub fn get_buffer(&self) -> Option<TextBuffer>;

    pub fn get_first(&self) -> u32;

    pub fn get_iter_at_line(&self, line: u32) -> TextIter;

    pub fn get_last(&self) -> u32;

    pub fn get_line_yrange(
        &self, 
        line: u32, 
        mode: GutterRendererAlignmentMode
    ) -> (i32, i32);

    pub fn get_view(&self) -> Option<TextView>;

    pub fn has_class(&self, line: u32, name: &str) -> bool;

    pub fn has_qclass(&self, line: u32, qname: Quark) -> bool;

    pub fn is_cursor(&self, line: u32) -> bool;

    pub fn is_prelit(&self, line: u32) -> bool;

    pub fn is_selected(&self, line: u32) -> bool;

    pub fn remove_class(&self, line: u32, name: &str);

    pub fn remove_qclass(&self, line: u32, qname: Quark);
}

Trait containing all GutterLines methods.

Feature: v5_0

Implementors

GutterLines

Required methods

pub fn add_class(&self, line: u32, name: &str)

Adds the class name to line.

name will be converted to a glib::Quark as part of this process. A faster version of this function is available via GutterLinesExt::add_qclass for situations where the glib::Quark is known ahead of time.

Feature: v5_0

line

a line number starting from zero

name

a class name

pub fn add_qclass(&self, line: u32, qname: Quark)

Adds the class denoted by qname to line.

You may check if a line has qname by calling GutterLinesExt::has_qclass.

You can remove qname by calling GutterLinesExt::remove_qclass.

Feature: v5_0

line

a line number starting from zero

qname

a class name as a glib::Quark

pub fn get_buffer(&self) -> Option<TextBuffer>

Gets the gtk::TextBuffer that the GutterLines represents.

Feature: v5_0

Returns

a gtk::TextBuffer

pub fn get_first(&self) -> u32

Gets the line number (starting from 0) for the first line that is user visible.

Feature: v5_0

Returns

a line number starting from 0

pub fn get_iter_at_line(&self, line: u32) -> TextIter

Gets a gtk::TextIter for the current buffer at line

Feature: v5_0

iter

a location for a gtk::TextIter

line

the line number

pub fn get_last(&self) -> u32

Gets the line number (starting from 0) for the last line that is user visible.

Feature: v5_0

Returns

a line number starting from 0

pub fn get_line_yrange(
&self,
line: u32,
mode: GutterRendererAlignmentMode
) -> (i32, i32)

Gets the Y range for a line based on mode.

The value for y is relative to the renderers widget coordinates.

Feature: v5_0

line

a line number starting from zero

mode

a GutterRendererAlignmentMode

y

a location for the Y position in widget coordinates

height

the line height based on mode

pub fn get_view(&self) -> Option<TextView>

Gets the gtk::TextView that the GutterLines represents.

Feature: v5_0

Returns

a gtk::TextView

pub fn has_class(&self, line: u32, name: &str) -> bool

Checks to see if GutterLinesExt::add_class was called with the name for line.

A faster version of this function is provided via GutterLinesExt::has_qclass for situations where the quark is known ahead of time.

Feature: v5_0

line

a line number starting from zero

name

a class name that may be converted, to a glib::Quark

Returns

true if line contains name

pub fn has_qclass(&self, line: u32, qname: Quark) -> bool

Checks to see if GutterLinesExt::add_qclass was called with the quark denoted by qname for line.

Feature: v5_0

line

a line number starting from zero

qname

a glib::Quark containing the class name

Returns

true if line contains qname

pub fn is_cursor(&self, line: u32) -> bool

Checks to see if line contains the insertion cursor.

Feature: v5_0

line

a line number starting from zero

Returns

true if the insertion cursor is on line

pub fn is_prelit(&self, line: u32) -> bool

Checks to see if line is marked as prelit. Generally, this means the mouse pointer is over the line within the gutter.

Feature: v5_0

line

a line number starting from zero

Returns

true if the line is prelit

pub fn is_selected(&self, line: u32) -> bool

Checks to see if the view had a selection and if that selection overlaps line in some way.

Feature: v5_0

line

a line number starting from zero

Returns

true if the line contains a selection

pub fn remove_class(&self, line: u32, name: &str)

Removes the class matching name from line.

A faster version of this function is available via GutterLinesExt::remove_qclass for situations where the glib::Quark is known ahead of time.

Feature: v5_0

line

a line number starting from zero

name

a class name

pub fn remove_qclass(&self, line: u32, qname: Quark)

Reverses a call to GutterLinesExt::add_qclass by removing the glib::Quark matching qname.

Feature: v5_0

line

a line number starting from zero

qname

a glib::Quark to remove from line

Loading content...