Trait sourceview5::prelude::GutterRendererPixbufExt

pub trait GutterRendererPixbufExt: 'static {
    pub fn get_gicon(&self) -> Option<Icon>;

    pub fn get_icon_name(&self) -> Option<GString>;

    pub fn get_paintable(&self) -> Option<Paintable>;

    pub fn get_pixbuf(&self) -> Option<Pixbuf>;

    pub fn overlay_paintable<P: IsA<Paintable>>(&self, paintable: &P);

    pub fn set_gicon<P: IsA<Icon>>(&self, icon: Option<&P>);

    pub fn set_icon_name(&self, icon_name: Option<&str>);

    pub fn set_paintable<P: IsA<Paintable>>(&self, paintable: Option<&P>);

    pub fn set_pixbuf(&self, pixbuf: Option<&Pixbuf>);

    pub fn get_property_paintable(&self) -> Option<Paintable>;

    pub fn connect_property_gicon_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_icon_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_paintable_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_pixbuf_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all GutterRendererPixbuf methods.

Implementors

GutterRendererPixbuf

Required methods

pub fn get_gicon(&self) -> Option<Icon>

Get the gicon of the renderer

Returns

a gio::Icon

pub fn get_icon_name(&self) -> Option<GString>

pub fn get_paintable(&self) -> Option<Paintable>

Gets a gdk::Paintable that was set with GutterRendererPixbufExt::set_paintable

Feature: v5_0

Returns

a gdk::Paintable or None

pub fn get_pixbuf(&self) -> Option<Pixbuf>

Get the pixbuf of the renderer.

Returns

a gdk_pixbuf::Pixbuf

pub fn overlay_paintable<P: IsA<Paintable>>(&self, paintable: &P)

Allows overlaying a paintable on top of any other image that has been set for the pixbuf. This will be applied when the widget is next snapshot.

Feature: v5_0

paintable

a gdk::Paintable

pub fn set_gicon<P: IsA<Icon>>(&self, icon: Option<&P>)

icon

the icon, or None.

pub fn set_icon_name(&self, icon_name: Option<&str>)

icon_name

the icon name, or None.

pub fn set_paintable<P: IsA<Paintable>>(&self, paintable: Option<&P>)

paintable

the paintable, or None.

pub fn set_pixbuf(&self, pixbuf: Option<&Pixbuf>)

pixbuf

the pixbuf, or None.

pub fn get_property_paintable(&self) -> Option<Paintable>

pub fn connect_property_gicon_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_icon_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_paintable_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_pixbuf_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...