Trait sourceview5::prelude::GutterRendererTextExt

pub trait GutterRendererTextExt: 'static {
    pub fn measure(&self, text: &str) -> (i32, i32);

    pub fn measure_markup(&self, markup: &str) -> (i32, i32);

    pub fn set_markup(&self, markup: &str);

    pub fn set_text(&self, text: &str);

    pub fn get_property_markup(&self) -> Option<GString>;

    pub fn get_property_text(&self) -> Option<GString>;

    pub fn connect_property_markup_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_text_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all GutterRendererText methods.

Implementors

GutterRendererText

Required methods

pub fn measure(&self, text: &str) -> (i32, i32)

Measures the text provided using the pango layout used by the GutterRendererText.

text

the text to measure.

width

location to store the width of the text in pixels, or None.

height

location to store the height of the text in pixels, or None.

pub fn measure_markup(&self, markup: &str) -> (i32, i32)

Measures the pango markup provided using the pango layout used by the GutterRendererText.

markup

the pango markup to measure.

width

location to store the width of the text in pixels, or None.

height

location to store the height of the text in pixels, or None.

pub fn set_markup(&self, markup: &str)

pub fn set_text(&self, text: &str)

pub fn get_property_markup(&self) -> Option<GString>

pub fn get_property_text(&self) -> Option<GString>

pub fn connect_property_markup_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_text_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...