Trait sourceview5::prelude::SnippetContextExt

pub trait SnippetContextExt: 'static {
    pub fn clear_variables(&self);

    pub fn expand(&self, input: &str) -> Option<GString>;

    pub fn get_variable(&self, key: &str) -> Option<GString>;

    pub fn set_constant(&self, key: &str, value: &str);

    pub fn set_line_prefix(&self, line_prefix: &str);

    pub fn set_tab_width(&self, tab_width: i32);

    pub fn set_use_spaces(&self, use_spaces: bool);

    pub fn set_variable(&self, key: &str, value: &str);

    pub fn connect_changed<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all SnippetContext methods.

Implementors

SnippetContext

Required methods

pub fn clear_variables(&self)

Removes all variables from the context.

Feature: v5_0

pub fn expand(&self, input: &str) -> Option<GString>

pub fn get_variable(&self, key: &str) -> Option<GString>

Gets the current value for a variable named key.

Feature: v5_0

key

the name of the variable

Returns

the value for the variable, or None

pub fn set_constant(&self, key: &str, value: &str)

Sets a constatnt within the context. This is similar to a variable set with SnippetContextExt::set_variable but is expected to not change during use of the snippet.

Examples would be the date or users name.

Feature: v5_0

key

the constant name

value

the value of the constant

pub fn set_line_prefix(&self, line_prefix: &str)

pub fn set_tab_width(&self, tab_width: i32)

pub fn set_use_spaces(&self, use_spaces: bool)

pub fn set_variable(&self, key: &str, value: &str)

Sets a variable within the context.

This variable may be overridden by future updates to the context.

Feature: v5_0

key

the variable name

value

the value for the variable

pub fn connect_changed<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

The "changed" signal is emitted when a change has been discovered in one of the chunks of the snippet which has caused a variable or other dynamic data within the context to have changed.

Feature: v5_0

Loading content...