Trait sourceview5::CompletionContextExt

pub trait CompletionContextExt: 'static {
    fn get_activation(&self) -> CompletionActivation;

    fn get_bounds(&self) -> Option<(TextIter, TextIter)>;

    fn get_buffer(&self) -> Option<Buffer>;

    fn get_busy(&self) -> bool;

    fn get_completion(&self) -> Option<Completion>;

    fn get_empty(&self) -> bool;

    fn get_language(&self) -> Option<Language>;

    fn get_start_iter(&self, iter: &mut TextIter);

    fn get_view(&self) -> Option<View>;

    fn get_word(&self) -> Option<GString>;

    fn set_proposals_for_provider<P: IsA<CompletionProvider>, Q: IsA<ListModel>>(
        &self, 
        provider: &P, 
        results: Option<&Q>
    );

    fn connect_property_busy_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_empty_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all CompletionContext methods.

Implementors

CompletionContext

Required methods

fn get_activation(&self) -> CompletionActivation

Gets the mode for which the context was activated.

Feature: v5_0

fn get_bounds(&self) -> Option<(TextIter, TextIter)>

Gets the bounds for the completion, which is the beginning of the current word (taking break characters into account) to the current insertion cursor.

If begin is non-None, it will be set to the start position of the current word being completed.

If end is non-None, it will be set to the insertion cursor for the current word being completed.

Feature: v5_0

begin

a gtk::TextIter

end

a gtk::TextIter

Returns

true if the marks are still valid and begin or end was set.

fn get_buffer(&self) -> Option<Buffer>

Gets the underlying buffer used by the context.

This is a convenience function to get the buffer via the Completion property.

Feature: v5_0

Returns

a gtk::TextBuffer or None

fn get_busy(&self) -> bool

Gets the "busy" property. This is set to true while the completion context is actively fetching proposals from registered CompletionProvider's.

Feature: v5_0

Returns

true if the context is busy

fn get_completion(&self) -> Option<Completion>

Gets the Completion that created the context.

Feature: v5_0

Returns

an Completion or None

fn get_empty(&self) -> bool

Checks if any proposals have been provided to the context.

Out of convenience, this function will return true if self is None.

Feature: v5_0

Returns

true if there are no proposals in the context

fn get_language(&self) -> Option<Language>

Gets the language of the underlying buffer, if any.

Feature: v5_0

Returns

a Language or None

fn get_start_iter(&self, iter: &mut TextIter)

fn get_view(&self) -> Option<View>

Gets the text view for the context.

Feature: v5_0

Returns

a View or None

fn get_word(&self) -> Option<GString>

Gets the word that is being completed up to the position of the insert mark.

Feature: v5_0

Returns

a string containing the current word

fn set_proposals_for_provider<P: IsA<CompletionProvider>, Q: IsA<ListModel>>(
&self,
provider: &P,
results: Option<&Q>
)

This function allows providers to update their results for a context outside of a call to CompletionProvider::populate_async. This can be used to immediately return results for a provider while it does additional asynchronous work. Doing so will allow the completions to update while the operation is in progress.

Feature: v5_0

provider

an CompletionProvider

results

a gio::ListModel or None

fn connect_property_busy_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_empty_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...