Trait sourceview5::CompletionProviderExt

pub trait CompletionProviderExt: 'static {
    fn activate<P: IsA<CompletionProposal>>(
        &self, 
        context: &CompletionContext, 
        proposal: &P
    );

    fn display<P: IsA<CompletionProposal>>(
        &self, 
        context: &CompletionContext, 
        proposal: &P, 
        cell: &CompletionCell
    );

    fn get_priority(&self, context: &CompletionContext) -> i32;

    fn get_title(&self) -> Option<GString>;

    fn is_trigger(&self, iter: &TextIter, ch: char) -> bool;

    fn key_activates<P: IsA<CompletionProposal>>(
        &self, 
        context: &CompletionContext, 
        proposal: &P, 
        keyval: u32, 
        state: ModifierType
    ) -> bool;

    fn list_alternates<P: IsA<CompletionProposal>>(
        &self, 
        context: &CompletionContext, 
        proposal: &P
    ) -> Vec<CompletionProposal>;

    fn populate_async<P: IsA<Cancellable>, Q: FnOnce(Result<ListModel, Error>) + Send + 'static>(
        &self, 
        context: &CompletionContext, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn populate_async_future(
        &self, 
        context: &CompletionContext
    ) -> Pin<Box_<dyn Future<Output = Result<ListModel, Error>> + 'static>>;

    fn refilter<P: IsA<ListModel>>(
        &self, 
        context: &CompletionContext, 
        model: &P
    );
}

Trait containing all CompletionProvider methods.

Implementors

CompletionProvider, CompletionSnippets, CompletionWords

Required methods

fn activate<P: IsA<CompletionProposal>>(
&self,
context: &CompletionContext,
proposal: &P
)

fn display<P: IsA<CompletionProposal>>(
&self,
context: &CompletionContext,
proposal: &P,
cell: &CompletionCell
)

fn get_priority(&self, context: &CompletionContext) -> i32

fn get_title(&self) -> Option<GString>

fn is_trigger(&self, iter: &TextIter, ch: char) -> bool

fn key_activates<P: IsA<CompletionProposal>>(
&self,
context: &CompletionContext,
proposal: &P,
keyval: u32,
state: ModifierType
) -> bool

fn list_alternates<P: IsA<CompletionProposal>>(
&self,
context: &CompletionContext,
proposal: &P
) -> Vec<CompletionProposal>

Providers should return a list of alternates to proposal or None if there are no alternates available. This can be used by the completion view to allow the user to move laterally through similar proposals, such as overrides of methods by the same name.

Feature: v5_0

context

a CompletionContext

proposal

a CompletionProposal

Returns

a glib::PtrArray of CompletionProposal or None.

fn populate_async<P: IsA<Cancellable>, Q: FnOnce(Result<ListModel, Error>) + Send + 'static>(
&self,
context: &CompletionContext,
cancellable: Option<&P>,
callback: Q
)

fn populate_async_future(
&self,
context: &CompletionContext
) -> Pin<Box_<dyn Future<Output = Result<ListModel, Error>> + 'static>>

fn refilter<P: IsA<ListModel>>(&self, context: &CompletionContext, model: &P)

Loading content...