Trait sourceview5::GutterExt

pub trait GutterExt: 'static {
    fn get_view(&self) -> Option<View>;

    fn insert<P: IsA<GutterRenderer>>(
        &self, 
        renderer: &P, 
        position: i32
    ) -> bool;

    fn remove<P: IsA<GutterRenderer>>(&self, renderer: &P);

    fn reorder<P: IsA<GutterRenderer>>(&self, renderer: &P, position: i32);

    fn get_property_window_type(&self) -> TextWindowType;
}

Trait containing all Gutter methods.

Implementors

Gutter

Required methods

fn get_view(&self) -> Option<View>

Returns

the associated View.

fn insert<P: IsA<GutterRenderer>>(&self, renderer: &P, position: i32) -> bool

Insert renderer into the gutter. If renderer is yet unowned then gutter claims its ownership. Otherwise just increases renderer's reference count. renderer cannot be already inserted to another gutter.

renderer

a gutter renderer (must inherit from GutterRenderer).

position

the renderer position.

Returns

true if operation succeeded. Otherwise false.

fn remove<P: IsA<GutterRenderer>>(&self, renderer: &P)

fn reorder<P: IsA<GutterRenderer>>(&self, renderer: &P, position: i32)

Reorders renderer in self to new position.

renderer

a gtk::CellRenderer.

position

the new renderer position.

fn get_property_window_type(&self) -> TextWindowType

The text window type on which the window is placed.

Loading content...