Trait sourceview5::SnippetChunkExt

pub trait SnippetChunkExt: 'static {
    fn copy(&self) -> Option<SnippetChunk>;

    fn get_context(&self) -> Option<SnippetContext>;

    fn get_focus_position(&self) -> i32;

    fn get_spec(&self) -> Option<GString>;

    fn get_text(&self) -> Option<GString>;

    fn get_text_set(&self) -> bool;

    fn get_tooltip_text(&self) -> Option<GString>;

    fn set_context<P: IsA<SnippetContext>>(&self, context: &P);

    fn set_focus_position(&self, focus_position: i32);

    fn set_spec(&self, spec: &str);

    fn set_text(&self, text: &str);

    fn set_text_set(&self, text_set: bool);

    fn set_tooltip_text(&self, tooltip_text: &str);

    fn get_property_context(&self) -> Option<SnippetContext>;

    fn get_property_focus_position(&self) -> i32;

    fn set_property_focus_position(&self, focus_position: i32);

    fn get_property_spec(&self) -> Option<GString>;

    fn set_property_spec(&self, spec: Option<&str>);

    fn get_property_text(&self) -> Option<GString>;

    fn set_property_text(&self, text: Option<&str>);

    fn get_property_text_set(&self) -> bool;

    fn set_property_text_set(&self, text_set: bool);

    fn connect_property_context_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_focus_position_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_spec_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_text_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_text_set_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_tooltip_text_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all SnippetChunk methods.

Implementors

SnippetChunk

Required methods

fn copy(&self) -> Option<SnippetChunk>

Copies the source snippet.

Feature: v5_0

Returns

A SnippetChunk

fn get_context(&self) -> Option<SnippetContext>

Gets the context for the snippet insertion.

Feature: v5_0

Returns

A SnippetContext

fn get_focus_position(&self) -> i32

Gets the SnippetChunk:focus-position.

The focus-position is used to determine how many tabs it takes for the snippet to advanced to this chunk.

A focus-position of zero will be the last focus position of the snippet and snippet editing ends when it has been reached.

A focus-position of -1 means the chunk cannot be focused by the user.

Feature: v5_0

Returns

the focus-position

fn get_spec(&self) -> Option<GString>

Gets the specification for the chunk.

The specification is evaluated for variables when other chunks are edited within the snippet context. If the user has changed the text, the SnippetChunk:text and SnippetChunk:text-set properties are updated.

Feature: v5_0

Returns

the specification, if any

fn get_text(&self) -> Option<GString>

Gets the SnippetChunk:text property.

The text property is updated when the user edits the text of the chunk. If it has not been edited, the SnippetChunk:spec property is returned.

Feature: v5_0

Returns

the text of the chunk

fn get_text_set(&self) -> bool

Gets the SnippetChunk:text-set property.

This is typically set when the user has edited a snippet chunk.

Feature: v5_0

fn get_tooltip_text(&self) -> Option<GString>

fn set_context<P: IsA<SnippetContext>>(&self, context: &P)

fn set_focus_position(&self, focus_position: i32)

Sets the SnippetChunk:focus-position property.

The focus-position is used to determine how many tabs it takes for the snippet to advanced to this chunk.

A focus-position of zero will be the last focus position of the snippet and snippet editing ends when it has been reached.

A focus-position of -1 means the chunk cannot be focused by the user.

Feature: v5_0

focus_position

the focus-position

fn set_spec(&self, spec: &str)

Sets the specification for the chunk.

The specification is evaluated for variables when other chunks are edited within the snippet context. If the user has changed the text, the SnippetChunk:text and SnippetChunk:text-set properties are updated.

Feature: v5_0

spec

the new specification for the chunk

fn set_text(&self, text: &str)

Sets the text for the snippet chunk.

This is usually used by the snippet engine to update the text, but may be useful when creating custom snippets to avoid expansion of any specification.

Feature: v5_0

text

the text of the property

fn set_text_set(&self, text_set: bool)

Sets the SnippetChunk:text-set property.

This is typically set when the user has edited a snippet chunk by the snippet engine.

Feature: v5_0

text_set

the property value

fn set_tooltip_text(&self, tooltip_text: &str)

fn get_property_context(&self) -> Option<SnippetContext>

fn get_property_focus_position(&self) -> i32

fn set_property_focus_position(&self, focus_position: i32)

fn get_property_spec(&self) -> Option<GString>

fn set_property_spec(&self, spec: Option<&str>)

fn get_property_text(&self) -> Option<GString>

fn set_property_text(&self, text: Option<&str>)

fn get_property_text_set(&self) -> bool

fn set_property_text_set(&self, text_set: bool)

fn connect_property_context_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_focus_position_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_spec_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_text_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_text_set_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_tooltip_text_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...