Trait sourceview5::SnippetExt

pub trait SnippetExt: 'static {
    pub fn add_chunk<P: IsA<SnippetChunk>>(&self, chunk: &P);

    pub fn copy(&self) -> Option<Snippet>;

    pub fn get_context(&self) -> Option<SnippetContext>;

    pub fn get_description(&self) -> Option<GString>;

    pub fn get_focus_position(&self) -> i32;

    pub fn get_language_id(&self) -> Option<GString>;

    pub fn get_n_chunks(&self) -> u32;

    pub fn get_name(&self) -> Option<GString>;

    pub fn get_nth_chunk(&self, nth: u32) -> Option<SnippetChunk>;

    pub fn get_trigger(&self) -> Option<GString>;

    pub fn set_description(&self, description: &str);

    pub fn set_language_id(&self, language_id: &str);

    pub fn set_name(&self, name: &str);

    pub fn set_trigger(&self, trigger: &str);

    pub fn get_property_buffer(&self) -> Option<TextBuffer>;

    pub fn get_property_description(&self) -> Option<GString>;

    pub fn set_property_description(&self, description: Option<&str>);

    pub fn get_property_focus_position(&self) -> i32;

    pub fn get_property_language_id(&self) -> Option<GString>;

    pub fn set_property_language_id(&self, language_id: Option<&str>);

    pub fn get_property_name(&self) -> Option<GString>;

    pub fn set_property_name(&self, name: Option<&str>);

    pub fn get_property_trigger(&self) -> Option<GString>;

    pub fn set_property_trigger(&self, trigger: Option<&str>);

    pub fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_description_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_focus_position_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_language_id_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    pub fn connect_property_trigger_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Snippet methods.

Feature: v5_0

Implementors

Snippet

Required methods

pub fn add_chunk<P: IsA<SnippetChunk>>(&self, chunk: &P)

Appends chunk to the self.

This may only be called before the snippet has been expanded.

Feature: v5_0

chunk

a SnippetChunk

pub fn copy(&self) -> Option<Snippet>

Does a deep copy of the snippet.

Feature: v5_0

Returns

A new Snippet

pub fn get_context(&self) -> Option<SnippetContext>

Gets the context used for expanding the snippet.

Feature: v5_0

Returns

an SnippetContext

pub fn get_description(&self) -> Option<GString>

Gets the description for the snippet.

Feature: v5_0

pub fn get_focus_position(&self) -> i32

Gets the current focus for the snippet. This is changed as the user tabs through focus locations.

Feature: v5_0

Returns

The focus position, or -1 if unset.

pub fn get_language_id(&self) -> Option<GString>

Gets the language-id used for the source snippet.

The language identifier should be one that matches a source language Language:id property.

Feature: v5_0

Returns

the language identifier

pub fn get_n_chunks(&self) -> u32

Gets the number of chunks in the snippet.

Note that not all chunks are editable.

Feature: v5_0

Returns

The number of chunks.

pub fn get_name(&self) -> Option<GString>

Gets the name for the snippet.

Feature: v5_0

pub fn get_nth_chunk(&self, nth: u32) -> Option<SnippetChunk>

Gets the chunk at nth.

Feature: v5_0

nth

the nth chunk to get

Returns

an SnippetChunk

pub fn get_trigger(&self) -> Option<GString>

Gets the trigger for the source snippet. A trigger is a word that can be expanded into the full snippet when the user presses Tab.

Feature: v5_0

Returns

A string or None

pub fn set_description(&self, description: &str)

Sets the description for the snippet.

Feature: v5_0

description

the snippet description

pub fn set_language_id(&self, language_id: &str)

Sets the language identifier for the snippet.

This should match the Language:id identifier.

Feature: v5_0

language_id

the language identifier for the snippet

pub fn set_name(&self, name: &str)

Sets the name for the snippet.

Feature: v5_0

name

the snippet name

pub fn set_trigger(&self, trigger: &str)

Sets the trigger for the snippet.

Feature: v5_0

trigger

the trigger word

pub fn get_property_buffer(&self) -> Option<TextBuffer>

pub fn get_property_description(&self) -> Option<GString>

pub fn set_property_description(&self, description: Option<&str>)

pub fn get_property_focus_position(&self) -> i32

pub fn get_property_language_id(&self) -> Option<GString>

pub fn set_property_language_id(&self, language_id: Option<&str>)

pub fn get_property_name(&self) -> Option<GString>

pub fn set_property_name(&self, name: Option<&str>)

pub fn get_property_trigger(&self) -> Option<GString>

pub fn set_property_trigger(&self, trigger: Option<&str>)

pub fn connect_property_buffer_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_description_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_focus_position_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_language_id_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_trigger_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...