Trait sourceview5::SnippetManagerExt

pub trait SnippetManagerExt: 'static {
    fn get_search_path(&self) -> Vec<GString>;

    fn get_snippet(
        &self, 
        group: Option<&str>, 
        language_id: Option<&str>, 
        trigger: &str
    ) -> Option<Snippet>;

    fn list_groups(&self) -> Vec<GString>;

    fn list_matching(
        &self, 
        group: Option<&str>, 
        language_id: Option<&str>, 
        trigger_prefix: Option<&str>
    ) -> Option<ListModel>;

    fn set_search_path(&self, dirs: &[&str]);

    fn connect_property_search_path_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all SnippetManager methods.

Feature: v5_0

Implementors

SnippetManager

Required methods

fn get_search_path(&self) -> Vec<GString>

Gets the list directories where self looks for snippet files.

Feature: v5_0

Returns

None-terminated array containing a list of snippet files directories. The array is owned by lm and must not be modified.

fn get_snippet(
&self,
group: Option<&str>,
language_id: Option<&str>,
trigger: &str
) -> Option<Snippet>

Queries the known snippets for the first matching group, language_id, and/or trigger. If group or language_id are None, they will be ignored.

Feature: v5_0

group

a group name or None

language_id

a Language:id or None

trigger

the trigger for the snippet

Returns

a Snippet or None if no matching snippet was found.

fn list_groups(&self) -> Vec<GString>

List all the known groups within the snippet manager.

The result should be freed with g_free, and the individual strings are owned by self and should never be freed by the caller.

Feature: v5_0

Returns

An array of strings which should be freed with g_free.

fn list_matching(
&self,
group: Option<&str>,
language_id: Option<&str>,
trigger_prefix: Option<&str>
) -> Option<ListModel>

Queries the known snippets for those matching group, language_id, and/or trigger_prefix. If any of these are None, they will be ignored when filtering the available snippets.

The gio::ListModel only contains information about the available snippets until gio::ListModel::get_item is called for a specific snippet. This helps reduce the number of gobject::Object's that are created at runtime to those needed by the calling application.

Feature: v5_0

group

a group name or None

language_id

a Language:id or None

trigger_prefix

a prefix for a trigger to activate

Returns

a gio::ListModel of Snippet.

fn set_search_path(&self, dirs: &[&str])

Sets the list of directories in which the GtkSourceSnippetManagerlooks for snippet files. If dirs is None, the search path is reset to default.

<note> <para> At the moment this function can be called only before the snippet files are loaded for the first time. In practice to set a custom search path for a SnippetManager, you have to call this function right after creating it. </para> </note>

Feature: v5_0

dirs

a None-terminated array of strings or None.

fn connect_property_search_path_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...