Trait sourceview5::StyleExt

pub trait StyleExt: 'static {
    pub fn apply<P: IsA<TextTag>>(&self, tag: &P);

    pub fn copy(&self) -> Option<Style>;

    pub fn get_property_background(&self) -> Option<GString>;

    pub fn get_property_background_set(&self) -> bool;

    pub fn get_property_bold(&self) -> bool;

    pub fn get_property_bold_set(&self) -> bool;

    pub fn get_property_foreground(&self) -> Option<GString>;

    pub fn get_property_foreground_set(&self) -> bool;

    pub fn get_property_italic(&self) -> bool;

    pub fn get_property_italic_set(&self) -> bool;

    pub fn get_property_line_background(&self) -> Option<GString>;

    pub fn get_property_line_background_set(&self) -> bool;

    pub fn get_property_pango_underline(&self) -> Underline;

    pub fn get_property_scale(&self) -> Option<GString>;

    pub fn get_property_scale_set(&self) -> bool;

    pub fn get_property_strikethrough(&self) -> bool;

    pub fn get_property_strikethrough_set(&self) -> bool;

    pub fn get_property_underline_color(&self) -> Option<GString>;

    pub fn get_property_underline_color_set(&self) -> bool;

    pub fn get_property_underline_set(&self) -> bool;
}

Trait containing all Style methods.

Implementors

Style

Required methods

pub fn apply<P: IsA<TextTag>>(&self, tag: &P)

This function modifies the gtk::TextTag properties that are related to the Style properties. Other gtk::TextTag properties are left untouched.

If self is non-None, applies self to tag.

If self is None, the related *-set properties of gtk::TextTag are set to false.

tag

a gtk::TextTag to apply styles to.

pub fn copy(&self) -> Option<Style>

Creates a copy of self, that is a new Style instance which has the same attributes set.

Returns

copy of self, call gobject::Object::unref when you are done with it.

pub fn get_property_background(&self) -> Option<GString>

pub fn get_property_background_set(&self) -> bool

pub fn get_property_bold(&self) -> bool

pub fn get_property_bold_set(&self) -> bool

pub fn get_property_foreground(&self) -> Option<GString>

pub fn get_property_foreground_set(&self) -> bool

pub fn get_property_italic(&self) -> bool

pub fn get_property_italic_set(&self) -> bool

pub fn get_property_line_background(&self) -> Option<GString>

pub fn get_property_line_background_set(&self) -> bool

pub fn get_property_pango_underline(&self) -> Underline

pub fn get_property_scale(&self) -> Option<GString>

pub fn get_property_scale_set(&self) -> bool

pub fn get_property_strikethrough(&self) -> bool

pub fn get_property_strikethrough_set(&self) -> bool

pub fn get_property_underline_color(&self) -> Option<GString>

pub fn get_property_underline_color_set(&self) -> bool

pub fn get_property_underline_set(&self) -> bool

Loading content...